Created Grid Map
I have created a map with a grid system. At the moment I have no idea on the size the map will be so I have chosen to allow it's dimensions to be assined in engine. As well as creating a dynamic map I have assigned the cells to hold references to their neighbours in a basic script.

Using this same script the neighbours can be accessed easily. This will make the pathfinding much easier.