Basic Pathfinding
I have implemented some basic apthfinging into the game. There is no visual elements in the pathfinding system but it works for the moment.

I am waiting on units and their moment before I can continue with the pathfinding elements.

Next I will probably sort out a resource script which will allow for the access of the three kinds of resources. It shouldn't take too long but it will give me some time to think about buildings and their placment on the cells as well as their movement blocking qualities.