Added Resources
I have added a basic resource tracker into the game. Right now there is only one (for the player) but a second could be added and given to the opponent (When they are created).

I have also added in adders and getters. This will be used to allow for the player to modify their resource value to build items.