Limited Unit Movement
I have added a limit on the unit movement within the game. The unit can only move a number of spaces up to their movement range, if they reach this number they will be unable to move again. I have also added a wait button which appears when the unit is selected, this button will use up all of the unit's movement points preventing them from moving again.

Next, I will add another unit/unit manager which will handle the opponent's units and add a combat system between them. This will also enable me to add in limits to prevent the player from doing anything during the opponents turn and not allow for the player to select multiple units and the opponent's units as well.