Attack And Destroy
I have added an attack command into the game, this will allow for the player to select and attack an opponent's unit. There is no visual cue to show when an enemy is selected, and there is no range limit to the attacking.

Next I will add some limitations to the attacking, and some visuals into the game to make it feel better and easier to use.