Attack Range limited
I have limited the attack range for the units to 1, this will need to be adjusted later for ranged units when they are added into the game but for now it works. What I have done is added a rsnge finder script which will check the cell neighbours for the opponent's unit's current cell and if it is in one of them then it will set a boolean inside unitStats to true allowing them to be selected by the player then, after they are selected it will use the legacy code to find the selected unit and run the attack code.

I still need to add some visuals into the game but for now the attacking code works quite well.

It might be time to introduce some of the AI code. I don't know yet, maybe it would be a goo idea to get all of the player code working first before trying to get an AI workng. But, it might be more fun to have an AI?