Added Colour
I have added colour onto the enemy selection process. This will make it a little easier to determine which enemy has beed selected. I will need to reset the colour and the selected enemy upon switching the selected unit. But for now its good.

Next, it would be a good idea to do a round of commenting the ingame code to make it easier to read.