Added Level Switching
I have added in int variable into the tile map manager class which will allow for the easy switching between levels within the game. as long as there is a level in the json file with a proper name then the level can be loaded.

Next, I will look over my code and do some commenting.