Cell Neighbours
I have created a script which will hold the list of adjacent cells each cell has. This system of neighbours will aid the new implementation of the path finding algorithm.

I think that I will tackle adding in units into the game next, this will start with movement in the game.