Added camera zooming
I have enabled the camera to zoom in and out on command, I have also added a min and max zoom value as to prevent the camera from being unusable.

Next, I should add a limit to the X and Y movement for the camera as well to prevent the player from leaving the map radius.