Updated Move Objects
I have updated my Move Objects scripts to allow for a position limiter into the scripts, this will prevent the connected object to be able to move endlessly in a single directio and away from the intended location in the game world.

I will next begin working on the Unit scripts within the game, starting with the positioning.