Basic Unit Movement
I have implemented a basic movement system into the game, this allows for the player to click on a cell and the units within the unit list will be moved to that selected cell.

I will next need to add a script which will check if the unit is an active unit, preventing multiple units from moving to the same cell.