Move Dist Limited
I have added in a pair of variables which will be used to limit the number of cells in the grid a unit can move. This is in the form of a max move value to limit the number of tiles which can be traversed and a current points spent, this will be used to keep track of the used points and compaered to the max points which can be spent. The way this is calculated is by finding the distance from the current cll to the desired cell and using the max points and the points already used to check if it is in range.

Following this, I will add in a second unit for the opponent and implement a basic turn system.