Limited Unit Movement
I have Limited the maximum distance the selected unit can travel as well as displaying their distance on the game map, this works by uisng the cell's neighbours and cheing if they are in range. I swapped to this rather tan looping through the full map to save on processing power.

Next, I will work on turn switching, and having a pair of units, one owned by the player the other by the opponent. and allowing each to move only on their own turn.