Radial Menu
Using a unity asset I have added in a radial menu into the game, I still need to figgure the ins' and outs of it but I have a circular 'Pie' Menu which centers with the unit at it's center.

Next, is to add the functionality of the buttons, I will start off with the movement button. This will allow for the prevention of moving while tying to press one of the other buttons, next I will implement the wait system, clearing all of that unit's actions.