Attack and Wait
I have added a basic attack system where when supplied with an attack target it will make the attack calculations and apply the attack. I have also added to the health management script to destroy the object with 0 health.

I have also added a simple wait system into the game which when activated reduces the current move points to maximum, preventing the unit for moving for the rest of the turn.