Added Targeting Script
I have allowed the player to find at target the opponent's units within the game. At the moment the scripts within the game are a little messy so I will need to clean them up a little, so thats which I intend to do next.

I have also added the first proper environment art into the game, it required a little resizing but I think it doesn't loom too bad. I also had to mess around with the camera settings but it fits I think.