Begun work on Ai turn management
I have started working on the AI's turn, I will an AI will have a static structure for controlling their units; Movement, Attacking, Building followed by Spawning new units. This will be their structure.

I have started with the movement side of their turn and I plan to find the AI unit's enemy target and move towards it, this will require me to evaluate the player unit's combat power uisng their stats as well as their distance to the unit the AI is controlling. I should start off with only a single unit on the player's side and add functionality into the player's unit_manager to calculate the combat powers for all of their units, allowing me to easily check them and add on the distance factor for easy exaluation.