Added obsticles and occupied booleans
I have added a pair of booleans into the game one to represent a cell which is cannot be moved to (tiles which are obsticles, these can be walls, mountains or in this case water) and a second cells which are occupied, these are cells which cannot be moved to because a unit is already there. This will help prevent unintended movement within the game.

Next, I plan to implement A* path finding into the game, but this is one big task ahead of me and will probably take a while to work as intended. At the moment I think i'll take a break and continue in a little while.