Added AI movement and Attacking
I have added a form of A* pathfinding into the game and at the moment the AI can choose a target and move towards that target. I have also changed the movement slightly; the new movement is a lerp between the old position and the new target position to make it feel a little better rather than instant movement to target location.

I have also made it so that if the AI unit starts it's turn adjacent to an opponent's unit it will attack it, I will need to make sure that if the unit ends it's turn next to an opponen't unit they will also attack. But for now the system works.

Next, I plan to add a health bar attached to each unit to easily display their current health while in the game for quality of life.