Added Health Bar
I have created a prefab for a basic unit within the game, using this prefab I have added a slider onto it which will act as it's health bar.

Next, I plan to add a spawning script I can add onto both the unit managers which will allow for me to spawn the different types of units into the game. This will allow for me to move onto the next step of the gameplay; the buildings.