Added Unit Spawning
I have added a spawning script which will be used to spawn 4 militia units onto both of the unit manager scripts. At the start of the game the units will be added into the game with a random set of stats and a random location, I will need to add a spawn point into the game to make sure the units spawn together.

Next I will need to make a HQ building for both players, this will allow for a win/lose condition to be programed into the game as well as a building block for the other building functions.

The HQ will also allow for me to designate spawn points for the starting units within the game (The 4 cells around the HQ, subject to change depending on cell size of HQ).