Created Info Bar
I have created a bar which will be used to hold the info for the game, at the moment I have created some placeholders for resource info and I have created a turn display, as well as updating it with the current turn player.

Next, I will create a HQ object for both of the players.