Begun working on buildings
I have begun working on implementing the buildings into the game, I have started off by uisng a placeholder for the HQ object. I have created a list to hold all of the buildings which will be spawned into the game.

I have also been adjusting the path finding algorithm. I will need to do a little more work on it though, storing old F-Scores and generating a new one to see if the new one is better than the old one giving it a root node it will probably require a fair bit of reworking to get it in a functional state when i mess things around.

I will also need to add into the game manager object a load order for to game so that when it starts critical objects are loaded before other objects attept to reference them.