Begun working on load order for the game
The load order for the game is important and to avoid any missing resources I have decided to make a start-up script for the game manager which will load all of the other items into the game. I have started with the turn manage and have moved onto the level manager. At this time I need to load into the game the sprites which will be used for the tiles on the map, I will use a similar method as loading the levels themselves all with their reference data.

To make all of this work I have had to move the main files into a Resources folder hence all of the changes, there were only a few files changed.