Sprite Loader/Map Loading
I have finished map loading with the addition of also setting the upper and lower bounds for the game camera. This is critical for the next step which is to load in the players, I will obviously start off with initilizing the player before the AI, and this will be a three step process; HQ/Building manager -> Units/Unit manager -> Resources and it's manager. If I follow this process it will prevent me accessing data before it has been loaded into the game.

I have also made a Sprite loader class which at the moment only handles the storing and loading of the Tiles for use in the map manager class but it has options to also be used with the units at a later time.