Setting up the player object
I started with creating a new object which will hold both the building and the unit managers. I then proceded to spawn the HQ using the HQ position stored on the level manager and usingthe HQ position's neighbours I spawned units around it.

Following that I started to link the user interface objects to their reference objects; I am in the process of linking the action radial menu's buttons to have the appropriate functions assigned.