Finished Load sequence for player object
I have finished connecting the components and resources which are required for the player objects to work within the game.

Next, I will make a start on the AI components to allow for the game to load properly every time.

Following that I will add in Unit to Base combat with the HQ objects allowing for a win/lose condition within the game.