Finished the load order.
I have added the AI onto the end of the load order, this will include them spawning the HQ and the units around it. I have also fully linked all of the data the AI requires to function.

Next, I will include the base combat and this will allow for me to create a build for the game, this build will be for testing purposes only and not for distribution. It will allow for me to find any bugs I may have missed and allow for me to fix them.