Added building combat
I have added building combat into the game, this comes with a game over screen as well. It will change depending upon which HQ was destroyed. However, the game doesn't have the functionality of AI targeting for the HQ but that will be the next step, and to prevent it from messing with the other path finding.