AI pathfinding to building and combat working
I have enabled the AI to find a path to either a building or a unit and begin combat with it, with that the game is technically playable. I will however need to add a lot more features to ensure that the game is in its desired state but for now the game works and can be tested for bugs/unintended functionality.

I will create a testing document in which the testing can be submitted. This will be the first build for the game and will most likely have issues and the fact of the lack of a title menu/navigation will make the game a little annoying to boot up and close. so that will be my next port of call.