Level Select plus Main Menu
I have added a main menu into the game which will allow for the player to both transition to the level select screen and exit the game. the level select screen will allow for players to select from a list of levels and then when the start level button is pressed will the proceed to load the level.

Next I will need to make a pause button for the game which will also allow for the player to end the game early and return to the main menu, I will also need to make adjustments to the game over screen to allow for the player to return to the level select screen.