Resources
I have begun the creation of the resources, however this requires me to also add in the buildings which produce said resources.  This is the main time taker, it requires me to add spawn loctaions into the level loader Json file which means adding the functionality and the testing associated with it all. after adding them though I will ensure that at the end of each turn the resource value will be addes to the current number and be updated on the screen.