Added resource generation
I have added resource generation into the game, this will produce resources at the end of each turn for the player, I still need to add one onto the AI but the one for the player works perfectly. I have also added that buildings require gold to produce and if a player runs out of gold the most recent buidlding will lose health. This will act as a balance to prevent the player from building too many buildings and to help to manage resources. I will need to also do this with food forthe units but that is a future problem.

Next I will add the resource gen/manufactury onto the Ai then link the player's resources onto the UI to give visual feedback on how much thier production is.