Connected Resources toUI
I have connected the player's resources to be displayed onto the UI, this will give the player an Idea on what they currently have and their progress.

Following this I have also given the AI their resource manager which will also increase from their buildings.

Next, I plan to add in a food cost on the Unit manager, this will allow for the units to use food each turn as a soft Unit cap for the players and if tehy run out of food then the weakest unit will lose health.