Player Unit spawning
I have added a spawn menu for new units within the game. thsi at the moment is an unlimited amount of spawning each turn as long as there is a free space next to the HQ. I plan to add a resource cos in next to limit this but for now the functionality of adding new units into the game works fine.

I have mainly focused on the HQ spaning units because it is limited to only militia and engineers. After I add in the unit spawn cost I will figgure out the building spawning forming using units for the game.

From what I have just done I belive that it will follow a similar structure as the unit spawning but I might need to use different menus and that might cause some confusion.