Comment Wave and Unit Spawn Cost
I have made a simple unit cost for the game, this will happen when a unit is spawned.

I have also commented a fair few files, in proper, easy to read style,  I haven't done all of them due to it taking a long time to go through each piece of data.

Next, I plan to implement a way to build new buildings. I also plan to update the radial menu for the units, after each action is completed hiding buttons as we go, for example if the unit is out of movement, the move button gets hidden. and if allof the actions for a unit is done, they cannot be selected again. this should help to make selecting and controlling units easier.