Added building
I have allowed allof the units to build buidings within the game, this is with a unit cost and a cell placement. The player will need to select a unit, select a building type and a cell beside the selected cell. This will then place the building within the game map.

This will need some testing but for now the building system is implemented.

Next I need to add a system to create engineers and then limit the building to only those units. I will then need to introduce the barracks building to create new types of units.

Following this I will need to design a combat menu for the game.