Limited Building to Engineers
I have limited the building functions to strictly engineers removing the other option from other units.

Next I will create a combat menu which will show both the player unit's stats as well as the opponents unit. This will allow for a combat decision to be made and will then allow for a animation window to be added at a later date.