Commenting and Added Todos
I have looked over the code and added a couple TODos for the game, this should help me to jump back into the coding for the game. I still have a few more files to comment and more Todos to add but doing this should help me to place the new functions in the correct place.