More comments
I have added a few more comments and an extra todo into the code. I have only a couple more file to comment and it will be all up to date with regions and proper comments, following this it will be much easier to find what needs doing (ctrl + f "Todo") and with the regions everything is nicely split into what it's purpose is.