Comments and Begun work on Combat Screen
I have finished off this wave of comments adding todos wherever appropriate.

I have started on the layout of the combat menu placing placeholder text where I would like the information to be displayed.  I have added a blank script onto the top level object for the menu and need to setup variables for the text element so that they can be changed within the game. The creation and assignemnt of the variables will be the next task.