Added combat menu functionality
I have added the combat menu for the game. This menu will appear when an enemy unit is selected for an attack, this will then display on the menu the stats of the player's unit, I still need to add the information for the AI's units as well as adding the combat breakdown in the middle. I also need to implement the cancel button for the combat.

Following this I will update the build and start working on some of the bugs and unintended features found during testing.