Combat Menu plus New Build
I have finsihed inputting the data into the combat menu. This will allow for a total combat breakdown before making an attack. I have finalised these changes and added them into the build folder updating this to 0.06.

I have a few scaling issues when it comes to the UI and should be re arranged so that they fit better in multiple screen sizes. I also need to create a in-game menu allaoing allowing for the return to the main menu mid-game.