Played around with UI scaling
I have adjusted the scale for the main menu items so that they are larger with bigger aspect ratios. This is the first step to buiding better UI for the game, next I plan to adjust the in-game UI before I build a pause menu for exiting and adjusting items mid game.