Begun work on Mini menu
I have started to work on the mini menu bu for whatever reason Unity seems to be messing up with a yet not before seen error for me; cannot reparent window to suggested parent. So I will need to look into that if a restart another day does't fix the issue.