Added Mini Menu
I have added a mini menu into the game which will pause the game and allow for the player to exit back to the main menu and start a new game.

Next, I will create a system to add unit sprites into the game and render them into the game world. I will probably follow a similar structure to the Tile map sprite loading system.