New Build
I have added in the new unit sprite into the game, this is just a place holder for now as it was designed for a different unit but its better to look at than a plain square.

Next, I plan to refine the AI's intelect by allowing for them to build new buildings and new units during the game.